+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - `tec`: Refactor project structure
+- `cli/aux/log`: Rename `TEC_LOG_V` into `TEC_LOG_I`
 
 ### Fixed